Fixed problem with analytics tracker id
