--- app/src/dev/java/com/jaspersoft/android/jaspermobile/JasperAnalytics.java
@@ -1,21 +1,0 @@
-package com.jaspersoft.android.jaspermobile;
-
-import android.content.Context;
-
-import com.google.android.gms.analytics.GoogleAnalytics;
-import com.google.android.gms.analytics.HitBuilders;
-import com.google.android.gms.analytics.Tracker;
-
-/**
- * @author Andrew Tivodar
- * @since 2.0
- */
-public class JasperAnalytics implements Analytics {
-    @Override
-    public void init(Context appContext) {
-    }
-
-    @Override
-    public void trackPrintEvent() {
-    }
-}
