+++ app/src/dev/res/xml/analytics_tracker.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <int name="ga_dispatchPeriod">15</int>
+    <bool name="ga_dryRun">true</bool>
+    <string name="ga_logLevel">verbose</string>
+
+</resources>
