+++ app/src/main/AndroidManifest.xml
@@ -60,6 +60,14 @@
             android:name="roboguice.annotations.packages"
             android:value="com.jaspersoft.android.jaspermobile" />
 
+        <meta-data
+            android:name="com.google.android.gms.analytics.globalConfigResource"
+            android:resource="@xml/analytics_tracker" />
+
+        <meta-data
+            android:name="com.google.android.gms.version"
+            android:value="@integer/google_play_services_version" />
+
         <!-- Main activity -->
         <activity
             android:name=".activities.navigation.NavigationActivity_">
@@ -217,6 +225,10 @@
         <service
             android:name=".network.XmlSpiceService"
             android:exported="false" />
+
+        <service android:name="com.google.android.gms.analytics.AnalyticsService"
+            android:enabled="true"
+            android:exported="false"/>
 
         <provider
             android:name=".db.MobileDbProvider"
