+++ app/src/main/java/com/jaspersoft/android/jaspermobile/Analytics.java
@@ -1,12 +1,17 @@
 package com.jaspersoft.android.jaspermobile;
 
-import android.content.Context;
+import android.app.Application;
 
 /**
  * @author Andrew Tivodar
  * @since 2.1
  */
 public interface Analytics {
-    void init(Context appContext);
-    void trackPrintEvent();
+    void init(Application appContext);
+    void trackPrintEvent(PrintType printType);
+
+    enum PrintType{
+        REPORT,
+        DASHBOARD
+    }
 }
