+++ app/src/main/java/com/jaspersoft/android/jaspermobile/activities/viewer/html/dashboard/BaseDashboardActivity.java
@@ -169,7 +169,7 @@ public abstract class BaseDashboardActivity extends RoboToolbarActivity
             onHomeAsUpCalled();
         }
         if (itemId == R.id.printAction) {
-            analytics.trackPrintEvent();
+            analytics.trackPrintEvent(Analytics.PrintType.DASHBOARD);
             ResourcePrintJob job = JasperPrintJobFactory.createDashboardPrintJob(webView, resource);
             JasperPrinter.print(job);
         }
