+++ app/src/production/AndroidManifest.xml
@@ -27,10 +27,6 @@
     <application>
 
         <meta-data
-            android:name="com.google.android.gms.version"
-            android:value="@integer/google_play_services_version" />
-
-        <meta-data
             android:name="com.crashlytics.ApiKey"
             android:value="${crashlyticsApiKey}" />
