--- app/src/production/res/xml/analytics_tracker.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <bool name="ga_autoActivityTracking">true</bool>
-    <integer name="ga_sessionTimeout">1800</integer>
-    <integer name="ga_dispatchPeriod">15</integer>
-
-    <screenName name="com.jaspersoft.android.jaspermobile.activities.navigation.NavigationActivity_">
-        Home screen
-    </screenName>
-
-    <screenName name="com.jaspersoft.android.jaspermobile.activities.inputcontrols.InputControlsActivity">
-        Report options screen
-    </screenName>
-
-    <screenName name="com.jaspersoft.android.jaspermobile.activities.viewer.html.report.ReportViewerActivity_">
-        Report viewer screen (visualize)
-    </screenName>
-
-    <screenName name="com.jaspersoft.android.jaspermobile.activities.viewer.html.report.ReportHtmlViewerActivity_">
-        Report viewer screen (REST)
-    </screenName>
-
-    <screenName name="com.jaspersoft.android.jaspermobile.activities.viewer.html.SavedReportHtmlViewerActivity_">
-        Saved html files screen
-    </screenName>
-
-    <screenName name="com.jaspersoft.android.jaspermobile.activities.viewer.html.Amber2DashboardActivity_">
-        Dashboard viewer screen (visualize)
-    </screenName>
-
-    <screenName name="com.jaspersoft.android.jaspermobile.activities.viewer.html.AmberDashboardActivity_">
-        Dashboard viewer screen (HTTP new)
-    </screenName>
-
-    <screenName name="com.jaspersoft.android.jaspermobile.activities.viewer.html.LegacyDashboardViewerActivity_">
-        Dashboard viewer screen (HTTP old)
-    </screenName>
-
-    <screenName name="com.jaspersoft.android.jaspermobile.activities.inputcontrols.SaveReportActivity_">
-        Saving report screen
-    </screenName>
-
-    <screenName name="com.jaspersoft.android.jaspermobile.activities.auth.AuthenticatorActivity">
-        Add account screen
-    </screenName>
-
-    <screenName name="com.jaspersoft.android.jaspermobile.activities.PrivacyPolicyActivity_">
-        Privacy Policy screen
-    </screenName>
-
-    <screenName name=".activities.settings.SettingsActivity_">
-        Settings screen
-    </screenName>
-
-    <string name="ga_trackingId">@string/google_analytics_tracking_id</string>
-</resources>
